% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_plots.R
\name{find_dist_slice}
\alias{find_dist_slice}
\title{returns positions in matrix which values are between val1 and val2}
\usage{
find_dist_slice(distM, val1, val2)
}
\description{
returns positions in matrix which values are between val1 and val2
}
