% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_plots.R
\name{find_recomb_names}
\alias{find_recomb_names}
\title{Prints names of sequences pairs which pairwise nucleotide distances lay between
val11-val12 in genomic region 1 and val21-val22 in genomic region 2/}
\usage{
find_recomb_names(distM1, val11, val12, distM2, val21, val22)
}
\description{
distM1 - distance matrix built for region 1
distM2 - distance matrix built for region 2
Prints strings "name1   name2   distance_in_region1   distance_in_region2"
}
