% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_plots.R
\name{get_distances}
\alias{get_distances}
\title{Returns dataframe with distances in between sequences in two genomic regions and pairwise distance plot}
\usage{
get_distances(dna_object, st1, e1, st2, e2)
}
\description{
Returns dataframe with distances in between sequences in two genomic regions and pairwise distance plot
}
