% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_plots.R
\name{plot_dist}
\alias{plot_dist}
\title{Plots pairwise distance comparison plots for different pairs of regions}
\usage{
plot_dist(dna_object, step, method, fig_dir, name_fig)
}
\description{
dna_object - list of DNA sequences (class DNAbin)
step - step for dna regions/ window for regions is equal to step
method - method of calculation distances ("pdist", "JC", "Kimura", "TN")
fig_dir - path to directory for saving figures
name_fig - prefix for figure name
created directory fig_dir/step and saves pairwise distance comparison plot there
}
