% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_plots.R
\name{plot_PDCP_control}
\alias{plot_PDCP_control}
\title{Control for Pairwise nucleotide Distance Correspondence Plot}
\usage{
plot_PDCP_control(dna_object, st1, e1, st2, e2)
}
\arguments{
\item{dna_object}{list of DNA sequences produced by `read.dna` function of `ape` package (`as.character = TRUE` mode)}

\item{st1}{start position of genome region 1}

\item{e1}{end position of genome region 1}

\item{st2}{start position of genome region 2}

\item{e2}{end position of genome region 2}
}
\value{
list with ggplot object with control plot for two regions, dataframe with distances in concatenated odd and even positions of alignment
}
\description{
Plots control for PDC plot.  Control plot visualizes the correspondence of p-distances in subalignments with odd and even positions of original alignment.
}
