% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_plots.R
\name{plot_rmse}
\alias{plot_rmse}
\title{plots heatmap with RMSE in pairwise distance comparison plot for each pair of genomic regions
dna_object -  list of DNA sequences (class DNAbin)
step
window - length of genomic regions to compare
method - method of calculation distances ("pdist", "JC", "Kimura", "TN")
modification - pairwise deletion of positions with gaps or not
returns matrix with rmse values for each pair f=of genomic regions}
\usage{
plot_rmse(dna_object, step, window, method, modification = NA)
}
\description{
plots heatmap with RMSE in pairwise distance comparison plot for each pair of genomic regions
dna_object -  list of DNA sequences (class DNAbin)
step
window - length of genomic regions to compare
method - method of calculation distances ("pdist", "JC", "Kimura", "TN")
modification - pairwise deletion of positions with gaps or not
returns matrix with rmse values for each pair f=of genomic regions
}
