% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_plots.R
\name{plot_PDCP}
\alias{plot_PDCP}
\title{Pairwise nucleotide Distance Correspondence Plot}
\usage{
plot_PDCP(dna_object, st1, e1, st2, e2)
}
\arguments{
\item{dna_object}{list of DNA sequences produced by `read.dna` function of `ape` package (`as.character = TRUE` mode)}

\item{st1}{start position of genome region 1}

\item{e1}{end position of genome region 1}

\item{st2}{start position of genome region 2}

\item{e2}{end position of genome region 2}
}
\value{
list with ggplot object with PDC plot for two regions, distance matrices for region 1 and 2
}
\description{
Each dot corresponds to a pair of nucleotide distances between
the same pair of genomes in two genomic regions with alignment positions `st1-e1` and `st2-e2` (see axis).
Returns list with ggplot, matrices of distances between pairs of sequences calculated for
`st1-e1` and `st2-e2` regions
}
