% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_plots.R
\name{plot_PDCP_control}
\alias{plot_PDCP_control}
\title{Function plots pairwise nucleotide distance comparison plot with control.}
\usage{
plot_PDCP_control(dna_object, st1, e1, st2, e2)
}
\value{
list with ggplot object with PDC and control plot for two regions, distance matrices for region 1 and 2
}
\description{
Each dot corresponds to a pair of nucleotide distances between
the same pair of genomes in two genomic regions - st1-e1 and st2-e2 (see axis).
Returns list with ggplot, matrices of distances between pairs of seqences calculated for
st1-e1 and st2-e2 regions
}
